classify_server/requirements.txt
werkzeug
numpy
pillow
torchvision
torch
flask
